Make app responsive for small and medium screens
- Refactor File/Folder structure following sazim principles.
- Fix localStorage bug
