+++ src/pages/HomePage/Components/EditForm/EditForm.jsx
@@ -1,3 +1,5 @@
+import { useState } from "react";
+
 import {
   Button,
   Center,
@@ -8,10 +10,9 @@ import {
   Title,
 } from "@mantine/core";
 import { DateInput } from "@mantine/dates";
-import { formatDueDate } from "../helper/helper";
-import { PRIORITY_OPTIONS } from "../constants/constant";
-import { COLORS } from "../constants/constant";
-import { useState } from "react";
+
+import { formatDueDate } from "../../HomePageHelpers";
+import { PRIORITY_OPTIONS, COLORS } from "../../HomePageConstants";
 
 const EditForm = ({ task, editTask, isOpen, closeModal }) => {
   const [editedTask, setEditedTask] = useState({ ...task });
@@ -50,8 +51,7 @@ const EditForm = ({ task, editTask, isOpen, closeModal }) => {
             onChange={(input) =>
               setEditedTask({ ...editedTask, dueDate: formatDueDate(input) })
             }
-            label="Date input"
-            placeholder="Date input"
+            label="Due Date"
           />
           <Center>
             <Button color={COLORS.BUTTON} type="submit">
